requirements.txt
websockets==12.0
numpy
requests
python-dotenv
